feat: don't rely on external modules for the accordion machine
