+++ packages/xstate-v5/machines/accordion/accordion.anatomy.ts
@@ -1,0 +1,4 @@
+import { createAnatomy } from "@zag-js/anatomy"
+
+export const anatomy = createAnatomy("accordion").parts("root", "item", "itemTrigger", "itemContent", "itemIndicator")
+export const parts = anatomy.build()
